o Added internal options for store intermediate data/file to RSP_DEBUG_PATH directory.
REALLY REALLY ODD: With R 3.0.1 patched (2013-09-12 r63919 on OSX and 2013-09-13 r63932 on Windows) 'R CMD build R.rsp' gives an error when building the package vignettes.  It turns out that in parseRaw() the 'text' string for constructing RspText(text) may be corrupt/contain NULs.  This is circumvented by a local function validateText(), which "touches" the string and the problem magically disappears.  Still don't know why though!

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-dots@1649 edb9625f-4e0d-4859-8d74-9fd3b1da38cb